Move appointment-approver resolution to composition layer
The previous commit relocated the appointmentApprovers helper into
@alga-psa/db to silence the no-feature-to-feature-imports lint error.
That was the wrong target: db is a base/horizontal layer and shouldn't
host feature-domain logic. The correct home is the composition layer.

Split the helper by layer instead of relocating it wholesale, since it
contained two distinct kinds of code:

- readApproverIdsFromConfig + ApproverConfigJson are pure parsing of
  scheduling's own availability_settings config_json. These move back
  into the scheduling feature package
  (packages/scheduling/src/lib/appointmentApprovers.ts). Its two
  consumers import them via a relative path (intra-feature).

- resolveAppointmentApproverUserIds + expandTeamsToUserIds are
  cross-feature: they join availability_settings + team_members + users.
  These move to msp-composition
  (packages/msp-composition/src/scheduling/appointmentApprovers.ts),
  which already depends on scheduling (composition -> feature is the
  correct direction). The composition module imports the pure parser
  from @alga-psa/scheduling/lib/appointmentApprovers.

Keeping the pure parser in scheduling avoids the dependency cycle that
would arise if scheduling had to import up from the composition layer.

Importers:
- client-portal action + server public appointment-request route now
  import resolveAppointmentApproverUserIds from
  @alga-psa/msp-composition/scheduling/appointmentApprovers.

Supporting changes:
- Revert the @alga-psa/db ./appointmentApprovers export and delete
  packages/db/src/lib/appointmentApprovers.ts.
- Re-add the ./lib/appointmentApprovers export to scheduling.
- Add a dedicated ./scheduling/appointmentApprovers subpath export to
  msp-composition so consumers don't pull in the client-component
  scheduling barrel.
- Add @alga-psa/msp-composition to client-portal dependencies (acyclic:
  msp-composition does not depend on client-portal).

Verified: eslint on the four touched files reports 0 errors and no
no-feature-to-feature-imports violation; msp-composition
packageDependencies test passes.
