+++ packages/client-portal/package.json
@@ -29,6 +29,7 @@
     "@alga-psa/core": "*",
     "@alga-psa/db": "*",
     "@alga-psa/formatting": "*",
+    "@alga-psa/msp-composition": "*",
     "@alga-psa/projects": "*",
     "@alga-psa/scheduling": "*",
     "@alga-psa/storage": "*",
