+++ packages/client-portal/src/actions/client-portal-actions/appointmentRequestActions.ts
@@ -31,7 +31,7 @@ import {
   getAvailableDates as getDatesFromService
 } from '../../services/availabilityService';
 import { createNotificationFromTemplateInternal } from '@alga-psa/notifications/actions';
-import { resolveAppointmentApproverUserIds } from '@alga-psa/db/appointmentApprovers';
+import { resolveAppointmentApproverUserIds } from '@alga-psa/msp-composition/scheduling/appointmentApprovers';
 import { isValidEmail } from '@alga-psa/core';
 import { isEnterprise } from '@alga-psa/core/features';
 import { format, type Locale } from 'date-fns';
