+++ packages/db/package.json
@@ -45,10 +45,6 @@
     "./workDate": {
       "import": "./dist/lib/workDate.js",
       "types": "./src/lib/workDate.ts"
-    },
-    "./appointmentApprovers": {
-      "import": "./dist/lib/appointmentApprovers.js",
-      "types": "./src/lib/appointmentApprovers.ts"
     }
   },
   "scripts": {
