+++ packages/msp-composition/package.json
@@ -14,6 +14,10 @@
       "import": "./src/scheduling/index.ts",
       "types": "./src/scheduling/index.ts"
     },
+    "./scheduling/appointmentApprovers": {
+      "import": "./src/scheduling/appointmentApprovers.ts",
+      "types": "./src/scheduling/appointmentApprovers.ts"
+    },
     "./scheduling/*": {
       "import": "./src/scheduling/*.tsx",
       "types": "./src/scheduling/*.tsx"
