+++ packages/scheduling/src/actions/appointmentRequestManagementActions.ts
@@ -38,7 +38,7 @@ import {
 import { generateICSBuffer, generateICSFilename, ICSEventData } from '../utils/icsGenerator';
 import { formatInTimeZone, fromZonedTime } from 'date-fns-tz';
 import { resolveTeamsMeetingService } from '../lib/teamsMeetingService';
-import { readApproverIdsFromConfig } from '@alga-psa/db/appointmentApprovers';
+import { readApproverIdsFromConfig } from '../lib/appointmentApprovers';
 
 export interface IAppointmentRequest {
   appointment_request_id: string;
